Carte terminate
- Aggiunte tutte le carte con le rispettive texture ed effetti su tutte le truppe

+++ Assets/CFX/Character/Character.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class Character : MonoBehaviour
 {
-    public byte hp, attack, defence;
+    public int hp, attack, defence;
 
     public virtual void Attack(Character target) 
     {
